code/test/requirements.txt
# requirements.txt
openai==1.30.1
pandas==2.1.3
python-dotenv==1.0.0
python-dateutil==2.8.2
regex==2023.10.3
